Proposal to allow Rule.check to identify the current rule
In markdown-it JS, via the `alt` list,
you could specify that a rule interrupted another block, only for certain block types.

This is a rough proposal to re-implement that feature,
such that an implementation of `Rule.check` could identify what the current rule was.

Currently, I'm just setting a string name on the `BlockState`, as this was the least intrusive way (i.e. does not require a change in the signature of check).
But perhaps there is a nicer way?